+++ docs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 6.0.2
 
 * Mappers definition enhancement for array indexes.
+* Bug fix: expect submapper of `"Foo"` to be `{".": "Foo"}`
 
 ## 6.0.1